{"types":"C","runs":43}

+++ problem18.java
@@ -3,8 +3,8 @@ import acm.program.ConsoleProgram;
 public class problem18 extends ConsoleProgram {
 	public void run(){
 		
-		double a = readDouble("Enter first number: ");
-		double b = readDouble("Enter second number: ");
+		int a = readInt("Enter first number: ");
+		int b = readInt("Enter second number: ");
 		
 		if (a>b) println(a);
 		if (b>a) println (b);
